+++ mne_bids/write.py
@@ -793,6 +793,7 @@ def make_dataset_description(path, name, data_license=None,
                              authors=None, acknowledgements=None,
                              how_to_acknowledge=None, funding=None,
                              references_and_links=None, doi=None,
+                             datasettype='raw',
                              verbose=False):
     """Create json for a dataset description.
 
@@ -828,11 +829,14 @@ def make_dataset_description(path, name, data_license=None,
         separated str like ['a', 'b', 'c'].
     doi : str | None
         The DOI for the dataset.
+    datasettype : str
+        Must be either "raw" or "derivative". Defaults to "raw".
+    verbose : bool
+        Set verbose output to True or False.
 
     Notes
     -----
     The required field BIDSVersion will be automatically filled by mne_bids.
-    The recommended field DatasetType will be automatically filled to be "raw".
 
     """
     # default author to make dataset description BIDS compliant
@@ -847,11 +851,13 @@ def make_dataset_description(path, name, data_license=None,
         funding = funding.split(', ')
     if isinstance(references_and_links, str):
         references_and_links = references_and_links.split(', ')
+    if datasettype not in ['raw', 'derivative']:
+        raise ValueError('`datasettype` must be either "raw" or "derivative."')
 
     fname = op.join(path, 'dataset_description.json')
     description = OrderedDict([('Name', name),
                                ('BIDSVersion', BIDS_VERSION),
-                               ('DatasetType', 'raw'),
+                               ('DatasetType', datasettype),
                                ('License', data_license),
                                ('Authors', authors),
                                ('Acknowledgements', acknowledgements),